apps/agent/inventory/requirements.txt
# Inventory Agent Dependencies

# Environment variables
python-dotenv>=1.0.0

# Supabase client
supabase>=2.0.0

# HTTP requests
requests>=2.28.0
